requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
Pillow==12.0.0
django-storages==1.14.6
python-decouple==3.8
django-filter==25.2